move SubscriptionMatcher to core module
