+++ client/pom.xml
@@ -50,12 +50,6 @@
             <version>${revision}${sha1}${changelist}</version>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>io.opencmw</groupId>
-            <artifactId>server</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
 
     </dependencies>
