+++ core/src/main/java/io/opencmw/utils/PathSubscriptionMatcher.java
@@ -1,4 +1,4 @@
-package io.opencmw.server;
+package io.opencmw.utils;
 
 import java.net.URI;
 import java.util.function.BiPredicate;
