+++ core/src/test/java/io/opencmw/utils/PathSubscriptionMatcherTest.java
@@ -1,4 +1,4 @@
-package io.opencmw.server;
+package io.opencmw.utils;
 
 import java.net.URI;
 import java.util.function.BiPredicate;
